+++ source/frontend/buttons/bbtakedown.js
@@ -1,9 +1,9 @@
 const { ActionRowBuilder, ButtonBuilder, ButtonStyle, MessageFlags, ComponentType, DiscordjsErrorCodes } = require('discord.js');
 const { ButtonWrapper } = require('../classes');
 const { SKIP_INTERACTION_HANDLING } = require('../../constants');
-const { getRankUpdates } = require('../shared');
+const { syncRankRoles } = require('../shared');
 
-/** @type {typeof import("../logic")} */
+/** @type {typeof import("../../logic")} */
 let logicLayer;
 
 const mainId = "bbtakedown";
@@ -37,8 +37,10 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				logicLayer.hunters.findOneHunter(interaction.user.id, interaction.guild.id).then(async hunter => {
 					hunter.decrement("xp");
 					const [season] = await logicLayer.seasons.findOrCreateCurrentSeason(interaction.guild.id);
-					logicLayer.seasons.changeSeasonXP(interaction.user.id, interaction.guildId, season.id, -1);
-					getRankUpdates(interaction.guild, logicLayer);
+					await logicLayer.seasons.changeSeasonXP(interaction.user.id, interaction.guildId, season.id, -1);
+					const descendingRanks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
+					const seasonUpdates = await logicLayer.seasons.updatePlacementsAndRanks(await logicLayer.seasons.getParticipationMap(season.id), descendingRanks);
+					syncRankRoles(seasonUpdates, descendingRanks, interaction.guild.id);
 				})
 
 				return collectedInteraction.reply({ content: "Your bounty has been taken down.", flags: MessageFlags.Ephemeral });