+++ source/frontend/commands/bounty/take-down.js
@@ -1,6 +1,6 @@
 const { ActionRowBuilder, StringSelectMenuBuilder, MessageFlags, ComponentType, DiscordjsErrorCodes } = require("discord.js");
 const { SubcommandWrapper } = require("../../classes");
-const { commandMention, getRankUpdates, bountiesToSelectOptions } = require("../../shared");
+const { commandMention, bountiesToSelectOptions, syncRankRoles } = require("../../shared");
 const { SKIP_INTERACTION_HANDLING } = require("../../../constants");
 
 module.exports = new SubcommandWrapper("take-down", "Take down one of your bounties without awarding XP (forfeit posting XP)",
@@ -34,8 +34,10 @@ module.exports = new SubcommandWrapper("take-down", "Take down one of your bount
 
 				hunter.decrement("xp");
 				const [season] = await logicLayer.seasons.findOrCreateCurrentSeason(interaction.guild.id);
-				logicLayer.seasons.changeSeasonXP(interaction.user.id, interaction.guildId, season.id, -1);
-				getRankUpdates(interaction.guild, logicLayer);
+				await logicLayer.seasons.changeSeasonXP(interaction.user.id, interaction.guildId, season.id, -1);
+				const descendingRanks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
+				const seasonUpdates = await logicLayer.seasons.updatePlacementsAndRanks(await logicLayer.seasons.getParticipationMap(season.id), descendingRanks);
+				syncRankRoles(seasonUpdates, descendingRanks, interaction.guild.members);
 
 				collectedInteraction.reply({ content: "Your bounty has been taken down.", flags: MessageFlags.Ephemeral });
 			}).catch(error => {