+++ source/frontend/commands/moderation/take-down.js
@@ -1,7 +1,7 @@
 const { ActionRowBuilder, StringSelectMenuBuilder, MessageFlags, ComponentType, DiscordjsErrorCodes } = require("discord.js");
 const { SubcommandWrapper } = require("../../classes");
 const { SAFE_DELIMITER, SKIP_INTERACTION_HANDLING } = require("../../../constants");
-const { getRankUpdates, bountiesToSelectOptions } = require("../../shared");
+const { bountiesToSelectOptions, syncRankRoles } = require("../../shared");
 
 module.exports = new SubcommandWrapper("take-down", "Take down another user's bounty",
 	async function executeSubcommand(interaction, runMode, ...[logicLayer]) {
@@ -42,8 +42,10 @@ module.exports = new SubcommandWrapper("take-down", "Take down another user's bo
 				logicLayer.hunters.findOneHunter(posterId, interaction.guild.id).then(async poster => {
 					poster.decrement("xp");
 					const [season] = await logicLayer.seasons.findOrCreateCurrentSeason(interaction.guildId);
-					logicLayer.seasons.changeSeasonXP(posterId, interaction.guildId, season.id, -1);
-					getRankUpdates(interaction.guild, logicLayer);
+					await logicLayer.seasons.changeSeasonXP(posterId, interaction.guildId, season.id, -1);
+					const descendingRanks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
+					const seasonUpdates = await logicLayer.seasons.updatePlacementsAndRanks(await logicLayer.seasons.getParticipationMap(season.id), descendingRanks);
+					syncRankRoles(seasonUpdates, descendingRanks, interaction.guild.members);
 				})
 				collectedInteraction.reply({ content: `<@${posterId}>'s bounty **${bounty.title}** has been taken down by ${interaction.member}.` });
 			});