+++ source/frontend/commands/moderation/xp-penalty.js
@@ -1,6 +1,6 @@
 const { MessageFlags } = require("discord.js");
 const { SubcommandWrapper } = require("../../classes");
-const { getRankUpdates } = require("../../shared");
+const { syncRankRoles } = require("../../shared");
 
 module.exports = new SubcommandWrapper("xp-penalty", "Reduce a bounty hunter's XP",
 	async function executeSubcommand(interaction, runMode, ...[logicLayer]) {
@@ -14,8 +14,10 @@ module.exports = new SubcommandWrapper("xp-penalty", "Reduce a bounty hunter's X
 		hunter.decrement({ xp: penaltyValue });
 		hunter.increment({ penaltyCount: 1, penaltyPointTotal: penaltyValue });
 		const [season] = await logicLayer.seasons.findOrCreateCurrentSeason(interaction.guildId);
-		logicLayer.seasons.changeSeasonXP(member.id, interaction.guildId, season.id, penaltyValue * -1);
-		getRankUpdates(interaction.guild, logicLayer);
+		await logicLayer.seasons.changeSeasonXP(member.id, interaction.guildId, season.id, penaltyValue * -1);
+		const descendingRanks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
+		const seasonUpdates = await logicLayer.seasons.updatePlacementsAndRanks(await logicLayer.seasons.getParticipationMap(season.id), descendingRanks);
+		syncRankRoles(seasonUpdates, descendingRanks, interaction.guild.members);
 		interaction.reply({ content: `<@${member.id}> has been penalized ${penaltyValue} XP.`, flags: MessageFlags.Ephemeral });
 		if (!member.user.bot) {
 			member.send(`You have been penalized ${penaltyValue} XP by ${interaction.member}. The reason provided was: ${interaction.options.getString("reason")}`);