+++ source/frontend/commands/raffle/index.js
@@ -7,9 +7,9 @@ let logicLayer;
 
 const mainId = "raffle";
 const { slashData: subcommandSlashData, executeDictionary: subcommandExecuteDictionary } = createSubcommandMappings(mainId, [
-	"./byrank.js",
-	"./bylevel.js",
-	"./announceupcoming.js"
+	"announce-upcoming.js",
+	"by-level.js",
+	"by-rank.js"
 ]);
 module.exports = new CommandWrapper(mainId, "Randomly select a bounty hunter from a variety of pools", PermissionFlagsBits.ManageGuild, false, [InteractionContextType.BotDM, InteractionContextType.Guild, InteractionContextType.PrivateChannel], 3000,
 	(interaction, runMode) => {