+++ source/frontend/commands/reset/index.js
@@ -7,8 +7,8 @@ let logicLayer;
 
 const mainId = "reset";
 const { slashData: subcommandSlashData, executeDictionary: subcommandExecuteDictionary } = createSubcommandMappings(mainId, [
-	"allhunterstats.js",
-	"serversettings.js"
+	"all-hunter-stats.js",
+	"server-settings.js"
 ]);
 module.exports = new CommandWrapper(mainId, "Reset all bounty hunter stats, bounties, or server configs", PermissionFlagsBits.ManageGuild, false, [InteractionContextType.Guild], 3000,
 	(interaction, runMode) => {